+++ karma.conf.js
@@ -9,7 +9,6 @@ module.exports = function(config) {
     // list of files / patterns to load in the browser
     files: [
       'vendor/npm/es6-shim/es6-shim.js',
-      'vendor/npm/es6-promise/dist/es6-promise.js',
       'vendor/npm/systemjs/dist/system.src.js',
       'test/test-main.js',