+++ public/app/plugins/datasource/elasticsearch/bucket_agg.js
@@ -77,7 +77,7 @@ function (angular, _, queryDef) {
       switch($scope.agg.type) {
         case 'terms': {
           settings.order = settings.order || "asc";
-          settings.size = settings.size || "0";
+          settings.size = settings.size || "10";
           settings.orderBy = settings.orderBy || "_term";
 
           if (settings.size !== '0') {